refactor(git-id-switcher): split shell.ts into types/csp/baseStyles/shell modules
shell.ts (432 lines, 4 concerns) is decomposed into focused modules:
- types.ts: SanitizedHtml brand, BodyClass union, VALID_BODY_CLASSES
- csp.ts: CspValidationError, format patterns, assertValidNonce/Lang,
  coerceLang, buildCspString
- baseStyles.ts: getBaseStyles, GITHUB_README_URL, focus-visible helpers
- shell.ts: buildHtmlShell + HtmlShellOptions only (142 lines)

Barrel (index.ts) re-exports updated in dependency order. Internal
imports in document/error/loading updated to import from their
authoritative source module. Test suite unchanged (barrel-based).

tsc clean, eslint clean, 100% statement coverage maintained.


🖥️ IDE: [VS Code](https://code.visualstudio.com/)
🔌 Extension: [Claude Code](https://claude.ai/download)

Model-Raw: claude-opus-4-6
